refactor: better semantics in the EC key inheritance
